requirements.txt
pandas
python-dotenv
sqlalchemy
firebirdsql
